Update readme
This updates the readme with an instruction to install vosk-cli directly from PyPI.
It also features an updated instruction to install from source. The recommended way of doing this is now `pip install .`, which will also install the dependencies from `requirements.txt`. Therefore the `Install dependencies` section now only features instructions to install ffmpeg.
The section about the language model download was slightly changed and now instructs the user to unzip their model in a `./models` directory instead of `usr/local...`, to reflect recent code changes. A fourth section was added to give a more detailed explanation on how to use vosk-cli.