Mapping Languages to Models
This patch improves the mapping of languages to models. This allows
users to install different models for the same language. Vosk-cli will
try matching languages in such a case.

If users want better matching they should use the `-m` option to
actually select the model they want to use anyway.